refactor: s3 list objects
- list bucketsと同様にリファクタリング
- errorをmain.goまで返す
