+++ pkg/service/service.go
@@ -1,4 +1,4 @@
-package s3
+package service
 
 import (
 	"github.com/aws/aws-sdk-go-v2/service/s3"
@@ -16,7 +16,6 @@ func (l *s3BucketsName) Set(v string) {
 
 // Receives a value of type ListBucketsOutput in the argument and returns bucket name in string.
 func SelectBucketName(input *s3.ListBucketsOutput) string {
-	// TODO: 引数のnullチェック入れる
 	ls := new(s3BucketsName)
 	for _, bucket := range input.Buckets {
 		ls.Set(*bucket.Name)
